Refactor translation.json files and mixins.pug
This commit refactors the translation.json files in the en and it directories, as well as the mixins.pug file. The changes include adding new sections for personal projects in both languages, updating project descriptions, and adding links to external resources. Additionally, the mixins.pug file has been modified to correctly render the project descriptions and links.